requirements.txt
    ffmpeg-python==0.2.0
    chardet==4.0.0
    dateutils==0.6.12
    idna==2.10
    lxml==4.6.3
    podgen==1.1.0
    python-dateutil==2.8.1
    pytz==2021.1
    requests==2.25.1
    six==1.15.0
    tinytag==1.8.1
    urllib3==1.26.4
    tqdm==4.64.1
